requirements.txt
dbt-postgres
dlt[postgres]
pandas
sqlalchemy

python-chess
stockfish
python-dotenv